requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
openai
python-dotenv
